[+FEATURE] TYPO3CR: Support node removal in workspaces
This adds support for staged removal of nodes: If a node is removed
in a workspace which does have a base workspace, the node is only
marked as removed, not actually removed. Once the workspace changes
are published, the node in the base workspace is really removed.

Resolves: #10810
Original-Commit-Hash: 067b6f4b11b9be981bce3a6e7ccea29cc37b51e2

+++ TYPO3.TYPO3CR/Classes/Domain/Model/Node.php
@@ -90,6 +90,15 @@ class Node {
 	protected $contentType = 'unstructured';
 
 	/**
+	 * If this is a removed node. This flag can and is only used in workspaces
+	 * which do have a base workspace. In a bottom level workspace nodes are
+	 * really removed, in other workspaces, removal is realized by this flag.
+	 *
+	 * @var boolean
+	 */
+	protected $removed = FALSE;
+
+	/**
 	 * @var \F3\TYPO3CR\Domain\Service\Context
 	 * @transient
 	 */
@@ -495,7 +504,22 @@ class Node {
 		foreach ($this->getChildNodes() as $childNode) {
 			$childNode->remove();
 		}
-		$this->nodeRepository->remove($this);
+
+		if ($this->workspace->getBaseWorkspace() === NULL) {
+			$this->nodeRepository->remove($this);
+		} else {
+			$this->removed = TRUE;
+		}
+	}
+
+	/**
+	 * If this node is a removed node.
+	 *
+	 * @return boolean
+	 * @author Robert Lemke <robert@typo3.org>
+	 */
+	public function isRemoved() {
+		return $this->removed;
 	}
 
 	/**
